Environment-based scene access
